+++ threadedcomments/migrations/0001_initial.py
@@ -17,7 +17,7 @@ class Migration(migrations.Migration):
             fields=[
                 ('comment_ptr', models.OneToOneField(parent_link=True, auto_created=True, primary_key=True, serialize=False, to='comments.Comment')),
                 ('title', models.TextField(verbose_name='Title', blank=True)),
-                ('tree_path', models.TextField(verbose_name='Tree path', editable=False, db_index=True)),
+                ('tree_path', models.CharField(verbose_name='Tree path', max_length=500, editable=False, db_index=True)),
                 ('last_child', models.ForeignKey(on_delete=django.db.models.deletion.SET_NULL, verbose_name='Last child', blank=True, to='threadedcomments.ThreadedComment', null=True)),
                 ('parent', models.ForeignKey(related_name='children', default=None, blank=True, to='threadedcomments.ThreadedComment', null=True, verbose_name='Parent')),
             ],