+++ threadedcomments/south_migrations/0001_initial.py
@@ -14,7 +14,7 @@ class Migration(SchemaMigration):
             ('title', self.gf('django.db.models.fields.TextField')(blank=True)),
             ('parent', self.gf('django.db.models.fields.related.ForeignKey')(default=None, related_name='children', null=True, blank=True, to=orm['threadedcomments.ThreadedComment'])),
             ('last_child', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['threadedcomments.ThreadedComment'], null=True, blank=True)),
-            ('tree_path', self.gf('django.db.models.fields.TextField')(db_index=True)),
+            ('tree_path', self.gf('django.db.models.fields.CharField')(db_index=True, max_length=500)),
         ))
         db.send_create_signal('threadedcomments', ['ThreadedComment'])
 
@@ -87,7 +87,7 @@ class Migration(SchemaMigration):
             'last_child': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['threadedcomments.ThreadedComment']", 'null': 'True', 'blank': 'True'}),
             'parent': ('django.db.models.fields.related.ForeignKey', [], {'default': 'None', 'related_name': "'children'", 'null': 'True', 'blank': 'True', 'to': "orm['threadedcomments.ThreadedComment']"}),
             'title': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
-            'tree_path': ('django.db.models.fields.TextField', [], {'db_index': 'True'})
+            'tree_path': ('django.db.models.fields.CharField', [], {'db_index': 'True', 'max_length': 500})
         }
     }